requirements.txt

streamlit
tensorflow-cpu==2.8.0
